Moves @check_value_lock to utils
